+++ src/components/features/explore/hooks/useExploreNFTs.ts
@@ -1,3 +1,4 @@
+import { convertIpfsToGatewayUrl } from '@/lib/ipfs-utils';
 import { EMOJI_CONTRACT_ADDRESS } from '@/lib/thirdweb';
 import { useContract, useContractRead } from '@thirdweb-dev/react';
 import { useEffect, useMemo, useState } from 'react';
@@ -23,18 +24,6 @@ const IPFS_GATEWAYS = ['https://ipfs.io/ipfs/'];
 
 // メタデータキャッシュ
 const metadataCache = new Map<string, NFTMetadata>();
-
-// IPFSのURLをゲートウェイURLに変換する関数
-const convertIpfsToGatewayUrl = async (ipfsUrl: string): Promise<string> => {
-  if (!ipfsUrl) return '';
-
-  if (ipfsUrl.startsWith('ipfs://')) {
-    const ipfsHash = ipfsUrl.replace('ipfs://', '');
-    // 最初のゲートウェイで試す
-    return `${IPFS_GATEWAYS[0]}${ipfsHash}`;
-  }
-  return ipfsUrl;
-};
 
 // メタデータを取得する関数（キャッシュ付き）
 const fetchMetadata = async (url: string): Promise<NFTMetadata> => {
