+++ src/lib/ipfs-utils.ts
@@ -1,0 +1,8 @@
+export const convertIpfsToGatewayUrl = (ipfsUrl: string): string => {
+  if (!ipfsUrl) return '';
+  if (ipfsUrl.startsWith('ipfs://')) {
+    const ipfsHash = ipfsUrl.replace('ipfs://', '');
+    return `https://ipfs.io/ipfs/${ipfsHash}`;
+  }
+  return ipfsUrl;
+};
